Remove some casts

+++ src/y21/d01.rs
@@ -5,18 +5,18 @@ use itertools::Itertools;
 crate::test::test_part!(test1, part1, 1162);
 crate::test::test_part!(test2, part2, 1190);
 
-pub fn part1() -> i64 {
+pub fn part1() -> usize {
 	read_integer_lines("input/2021/01.txt")
 		.tuple_windows()
 		.filter(|(prev, next)| prev < next)
-		.count() as i64
+		.count()
 }
 
-pub fn part2() -> i64 {
+pub fn part2() -> usize {
 	read_integer_lines("input/2021/01.txt")
 		.tuple_windows()
 		.map(|(first, second, third)| first + second + third)
 		.tuple_windows()
 		.filter(|(prev, next)| prev < next)
-		.count() as i64
+		.count()
 }
